Update plaid.ex
I like your solution. I modified the setting of the `cred` local variable to be more explicit (grabbing `client_id` and `secret` as opposed to deleting unwanted variables. I also modified the concatenation of options using `Keyword.merge` to ensure that if options are provided in `config.exs`, then they are overridden when passed in the `config` argument or `options` argument in `Plaid.make_request_with_cred/6`. The priority is `options` argument -> `config` argument -> `config.exs`.