Add support for personal_finance_category
Plaid encourages all implementations to set the
`include_personal_finance_category` option to `true` and use the new
`personal_finance_category` field instead of the old `category` field,
as it is intended to provide higher accuracy and more meaningful
categories.

Implement the encoder for `PersonalFinanceCategory` and use it when
decoding  responses from both `Transactions.get` and
`Transactions.sync`. If the `include_personal_finance_category` option
is not set in the request, the field remains `nil`, otherwise is is
populated with the map containing the primary and detailed categories.